Refactor to share a good chunk of Keycloak logic across apps
