+++ src/filing/App.jsx
@@ -7,7 +7,8 @@ import Footer from './common/Footer.jsx'
 import BrowserBlocker from './common/BrowserBlocker.jsx'
 import Loading from '../common/LoadingIcon.jsx'
 import * as AccessToken from '../common/api/AccessToken.js'
-import { getKeycloak, refresh, login } from './utils/keycloak.js'
+import { refresh, login } from './utils/keycloak.js'
+import { getKeycloak } from '../common/api/Keycloak.js'
 import isRedirecting from './actions/isRedirecting.js'
 import updateFilingPeriod from './actions/updateFilingPeriod.js'
 import { detect } from 'detect-browser'
