+++ src/filing/index.js
@@ -7,24 +7,17 @@ import { Provider } from 'react-redux'
 import thunkMiddleware from 'redux-thunk'
 
 import { Switch, Route, Redirect } from 'react-router-dom'
-import Keycloak from 'keycloak-js'
 import AppContainer from './App.jsx'
 import HomeContainer from './home/container.jsx'
 import InstitutionContainer from './institutions/container.jsx'
 import SubmissionRouter from './submission/router.jsx'
-import { setKeycloak, mockKeycloak } from './utils/keycloak.js'
+import { initKeycloak } from '../common/api/Keycloak'
 import { setStore } from './utils/store.js'
 import appReducer from './reducers'
 import { withAppContext } from '../common/appContextHOC'
 
+initKeycloak()
 const middleware = [thunkMiddleware]
-if(process.env.REACT_APP_ENVIRONMENT === 'CI') 
-  setKeycloak(mockKeycloak)
-else if(process.env.NODE_ENV === 'development') {
-  setKeycloak(Keycloak(process.env.PUBLIC_URL + '/local_keycloak.json'))
-} else {
-  setKeycloak(Keycloak(process.env.PUBLIC_URL + '/keycloak.json'))
-}
 
 let store
 if (process.env.NODE_ENV !== 'production') {
